Fix: CSS + Logica do preço vir com centavos

+++ BackEnd/src/database/migrations/202210171954-Product.js
@@ -15,7 +15,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(9,2)
       },
       description: {
         allowNull: false,
